requirements.txt
-i https://pypi.org/simple
astroid==3.0.2; python_full_version >= '3.8.0'
bcrypt==4.1.2; python_version >= '3.7'
black==23.12.1; python_version >= '3.8'
cffi==1.16.0; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
cryptography==41.0.7; python_version >= '3.7'
dill==0.3.7; python_version >= '3.11'
iniconfig==2.0.0; python_version >= '3.7'
isort==5.13.2; python_full_version >= '3.8.0'
mccabe==0.7.0; python_version >= '3.6'
mypy-extensions==1.0.0; python_version >= '3.5'
packaging==23.2; python_version >= '3.7'
paramiko==3.4.0; python_version >= '3.6'
pathspec==0.12.1; python_version >= '3.8'
phoenix6==24.0.0b5; python_version >= '3.7' and python_version < '4'
photonlibpy==2024.1.1b42
pint==0.23; python_version >= '3.9'
platformdirs==4.1.0; python_version >= '3.8'
pluggy==1.3.0; python_version >= '3.8'
pycparser==2.21
pyfrc==2024.0.0b3; platform_machine != 'roborio' and platform_machine != 'armv7l' and platform_machine != 'aarch64'
pylint==3.0.3; python_full_version >= '3.8.0'
pynacl==1.5.0; python_version >= '3.6'
pynetconsole==2.0.4; python_version >= '3.6'
pyntcore==2024.0.0b4; python_version >= '3.8'
pytest==7.4.4; python_version >= '3.7'
pytest-reraise==2.1.2; python_full_version >= '3.6.1' and python_full_version < '4.0.0'
robotpy[all]==2024.0.0b4; python_version < '3.13' and python_version >= '3.8'
robotpy-apriltag==2024.0.0b4
robotpy-cli==2024.0.0; python_version >= '3.8'
robotpy-commands-v2==2024.0.0b3; python_version >= '3.8'
robotpy-cscore==2024.0.0b4
robotpy-ctre==2024.0.0b4
robotpy-hal==2024.0.0b4; python_version >= '3.8'
robotpy-halsim-ds-socket==2024.0.0b4
robotpy-halsim-gui==2024.0.0b4; platform_machine != 'roborio' and platform_machine != 'armv7l' and platform_machine != 'aarch64'
robotpy-halsim-ws==2024.0.0b4
robotpy-installer==2024.0.2; platform_machine != 'roborio' and platform_machine != 'armv7l' and platform_machine != 'aarch64'
robotpy-navx==2024.0.0b1
robotpy-pathplannerlib==2024.0.0b6.post3; python_version >= '3.8'
robotpy-rev==2024.0.0b1.post1
robotpy-wpilib-utilities==2024.0.0b1; python_version >= '3.7'
robotpy-wpimath==2024.0.0b4; python_version >= '3.8'
robotpy-wpinet==2024.0.0b4; python_version >= '3.8'
robotpy-wpiutil==2024.0.0b4; python_version >= '3.8'
setuptools==69.0.3; python_version >= '3.8'
tomli==2.0.1; python_version >= '3.7'
tomlkit==0.12.3; python_version >= '3.7'
typing-extensions==4.9.0; python_version >= '3.8'
wpilib==2024.0.0b4; python_version >= '3.8'
